Move XmlErrorResponse
