+++ src/ops/create_bucket.rs
@@ -1,8 +1,8 @@
 //! [`CreateBucket`](https://docs.aws.amazon.com/AmazonS3/latest/API/API_CreateBucket.html)
 
-use crate::error::S3Result;
+use crate::error::{S3Result, XmlErrorResponse};
 use crate::error_code::S3ErrorCode;
-use crate::output::{wrap_output, S3Output, XmlErrorResponse};
+use crate::output::{wrap_output, S3Output};
 use crate::utils::{deserialize_xml_body, RequestExt, ResponseExt};
 use crate::{Body, BoxStdError, Request, Response};
