+++ src/ops/head_bucket.rs
@@ -1,8 +1,8 @@
 //! [`HeadBucket`](https://docs.aws.amazon.com/AmazonS3/latest/API/API_HeadBucket.html)
 
-use crate::error::S3Result;
+use crate::error::{S3Result, XmlErrorResponse};
 use crate::error_code::S3ErrorCode;
-use crate::output::{S3Output, XmlErrorResponse};
+use crate::output::S3Output;
 use crate::utils::Apply;
 use crate::{Body, BoxStdError, Response};
