+++ src/ops/head_object.rs
@@ -1,8 +1,8 @@
 //! [`HeadObject`](https://docs.aws.amazon.com/AmazonS3/latest/API/API_HeadObject.html)
 
-use crate::error::S3Result;
+use crate::error::{S3Result, XmlErrorResponse};
 use crate::error_code::S3ErrorCode;
-use crate::output::{wrap_output, S3Output, XmlErrorResponse};
+use crate::output::{wrap_output, S3Output};
 use crate::utils::{time, RequestExt, ResponseExt};
 use crate::{BoxStdError, Request, Response};
