+++ src/ops/list_objects_v2.rs
@@ -1,9 +1,9 @@
 //! [`ListObjectsV2`](https://docs.aws.amazon.com/AmazonS3/latest/API/API_ListObjectsV2.html)
 
-use crate::error::S3Result;
+use crate::error::{S3Result, XmlErrorResponse};
 use crate::error_code::S3ErrorCode;
 use crate::headers::names::X_AMZ_REQUEST_PAYER;
-use crate::output::{wrap_output, S3Output, XmlErrorResponse};
+use crate::output::{wrap_output, S3Output};
 use crate::query::GetQuery;
 use crate::utils::{RequestExt, ResponseExt, XmlWriterExt};
 use crate::{BoxStdError, Request, Response};
